net/{stress,virt}: Migrate (partly) to the new API
Migrate tests using virt_lib.sh and ipsec_lib.sh.
It was needed to migrate it all in once, as macsec0{1,2}.sh tests are
using both libraries.

This includes processing getopts parameters and help with API helpers
(TST_OPTS, TST_PARSE_ARGS and TST_USAGE) as this is a proper way of
handling parameters by API.

Signed-off-by: Petr Vorel <pvorel@suse.cz>