net/ipsec: Fix getopts parsing -s parameter
legacy API in tst_net.sh is not able to detect quoted parameters,
because input parameters for getopts are passing as '$*' to tst_net_read_opts.
This drops quotes so only first number from -s is taken.
The solution is to use ':' as array separator.

NOTE: Although this could be easily fixed by calling getopts directly,
the same problem will reappear when tests will be migrated to the new
API (input parameters for getopts are in tst_test.sh passed via TST_ARGS
variable).

Fixes: 89a83d8a0 ("net/ipsec_lib.sh: Process getopts parameters with
TST_OPTS and TST_PARSE_ARGS")

Signed-off-by: Petr Vorel <pvorel@suse.cz>