OVL_MNT: put overlayfs lower, upper, work, mnt dir in separated mountpoint
Some tests are mounting overlayfs internally and run tests on it.
This mount can fail if the filesystem we are running on does not
support overlay mount upon it. For example, we are already running
tests on overlayfs or NFS, or CIFS. Test will report broken and
failure.

Fixing this by put overlayfs dirs in a separaed mountpoint, like
readahead02 by Amir.

Signed-off-by: Xiong Zhou <xzhou@redhat.com>
Reviewed-by: Li Wang <liwang@redhat.com>