shell: Introduce TST_TIMEOUT variable
to unify shell API with C API.

TST_TIMEOUT should be set in tests only, it's equivalent of
tst_test->timeout in C API.

+ add checks requiring both TST_TIMEOUT and LTP_TIMEOUT_MUL >= 1,
that allow to set TST_TIMEOUT lower than the default 300 sec
(might be useful for some case).
LTP_TIMEOUT_MUL can be float, but that adds awk as a dependency.

+ replace LTP_TIMEOUT_MUL in memcg_stress_test.sh with TST_TIMEOUT.

+ document both variables.

Suggested-by: Clemens Famulla-Conrad <cfamullaconrad@suse.de>
Signed-off-by: Petr Vorel <pvorel@suse.cz>