shell: Introduce LTP_TIMEOUT variable
to unify shell API with C API.

LTP_TIMEOUT should be set in tests only, it's equivalent of
tst_test->timeout in C API.

+ add checks requiring both LTP_TIMEOUT and LTP_TIMEOUT_MUL >= 1,
that allow to set LTP_TIMEOUT lower than the default 300 sec
(might be useful for some case).
LTP_TIMEOUT_MUL can be float, but that adds awk as a dependency.

+ replace LTP_TIMEOUT_MUL in memcg_stress_test.sh with LTP_TIMEOUT.

+ document both variables.

Suggested-by: Clemens Famulla-Conrad <cfamullaconrad@suse.de>
Signed-off-by: Petr Vorel <pvorel@suse.cz>