open_posix/pi_test*: Remove all tests
There are several reasons, why these tests should be removed.

Most important: They do not really test anything. The result of these
tests has to be inspected manually do check, if the behavior is correct,
but the manual inspection using the given tools is not possible, because
these tests were written for systems much slower than todays systems.

The problems are described in more detail in
https://github.com/linux-test-project/ltp/issues/584

Reviewed-by: Petr Vorel <pvorel@suse.cz>
Acked-by: Cyril Hrubis <chrubis@suse.cz>
Acked-by: Jan Stancek <jstancek@redhat.com>
Signed-off-by: Joerg Vehlow <joerg.vehlow@aox-tech.de>
Signed-off-by: Petr Vorel <pvorel@suse.cz>